+++ wiki/wiki_md/about/changelog.md
@@ -15,6 +15,7 @@ This page documents the changes and fixes made in Quantum Things, compared to th
 - Added ore dictionary support for various items and crafting recipes, improving compatibility with other mods and modpacks.
 - Added the ability to configure whether Spectre Tools should be disabled.
 - Added the ability to configure whether Spectre Armor should be disabled.
+- Added the ability to configure whether Luminous Blocks should emit light (disabled by default).
 
 ### Fixes